+++ docs/Changelog.md
@@ -13,6 +13,7 @@
 * [#3997](https://github.com/TouK/nussknacker/pull/3997) Removal of obsolete `subprocessVersions`.
 * [#4060](https://github.com/TouK/nussknacker/pull/4060) Notifications about pending deployments of other user not appear anymore.
 * [#4071](https://github.com/TouK/nussknacker/pull/4071) Change BCrypt library
+* [#4076](https://github.com/TouK/nussknacker/pull/4076) Fix for respecting interpreter timeout in Lite
 
 1.8.1 (28 Feb 2023)
 ------------------------